Working example of gropuing git buttons.
